requirements.txt
beautifulsoup4==4.6.3
peewee==3.7.1
pymystem3==0.2.0
python-telegram-bot==11.1.0